requirements.txt
fastapi
uvicorn
requests
supabase
python-dotenv
pydantic
